Draw one out of two faces
This is a hard-coded solution to the problem given to
me by my mentor Guillaume Damiand. Here, we construct
only one out of every two face by inverting a bool variable
at every iteration of the 'for' loop which aids us in
drawing our faces). This solution can be extended to edges
and vertices as well.

We are terming this solution as 'hard-coded' because it will only
work for the examples in 'Polyhedron'. The next step is to generalise
this solution by modifying 'face_begin()' function.